chore: refactor main.py and clean more comments
